refactor: extract admin debug projection
