+++ crates/dcc-mcp-gateway/src/gateway/admin/mod.rs
@@ -25,7 +25,8 @@
 //!
 //! The runtime UI is a single inline HTML string (`admin/html.rs`). The
 //! `dcc-mcp-gateway-admin` crate owns the trace domain, compact, link,
-//! issue-report, statistics, analytics, governance, artifact, activity, task-outcome, and traffic projections,
+//! issue-report, statistics, analytics, governance, artifact, activity, task-outcome,
+//! debug-bundle, postmortem, and traffic projections,
 //! Vite/npm build boundary, and embedded asset; this module owns the
 //! gateway-specific data-source/HTTP adapters and API handlers.
 //!
